7/16 회의 종합 완료 : sampleMap완성 -> imgae클리핑까지
해야될 것
권찬미 : 인벤토리 구현
이현호 : 충돌 구현
신진경 : 메뉴 클래스 구현( UI에 추가 )
허정훈 : 맵매니저 구현

+++ Pokemon_Mystery_Dungeon/dImage.cpp
@@ -15,7 +15,7 @@ dImage::~dImage()
 {
 }
 
-HRESULT dImage::init(LPCWCHAR fileName, float width, float height)
+HRESULT dImage::init(LPCWCHAR fileName, int width, int height)
 {
 	if (_imageInfo != NULL) release();
 
@@ -66,7 +66,7 @@ HRESULT dImage::init(LPCWCHAR fileName, float width, float height)
 	return S_OK;
 }
 
-HRESULT dImage::init(LPCWCHAR fileName, float width, float height, int frameX, int frameY)
+HRESULT dImage::init(LPCWCHAR fileName, int width, int height, int frameX, int frameY)
 {
 	if (_imageInfo != NULL) release();
 
@@ -154,7 +154,7 @@ void dImage::render(int destX, int destY, float opacity)
 	DTDMANAGER->getBitRenderTarget()->DrawBitmap(_imageInfo->bitMap, destRect, opacity);
 }
 
-void dImage::render(int destX, int destY, float sourX, float sourY, float sourWidth, float sourHeight, float opacity)
+void dImage::render(int destX, int destY, int sourX, int sourY, int sourWidth, int sourHeight, float opacity)
 {
 	D2D1_RECT_F destRect = dRectMake(destX, destY, sourWidth, sourHeight);
 	D2D1_RECT_F sourRect = dRectMake(sourX, sourY, sourWidth, sourHeight);
